sig: rename sigconfig package to config
