+++ go/sig/BUILD.bazel
@@ -16,7 +16,7 @@ go_library(
         "//go/lib/serrors:go_default_library",
         "//go/lib/sigdisp:go_default_library",
         "//go/lib/sigjson:go_default_library",
-        "//go/pkg/sig/sigconfig:go_default_library",
+        "//go/pkg/sig/config:go_default_library",
         "//go/sig/egress:go_default_library",
         "//go/sig/internal/base:go_default_library",
         "//go/sig/internal/ingress:go_default_library",
