+++ go/sig/internal/sigcmn/BUILD.bazel
@@ -17,7 +17,7 @@ go_library(
         "//go/lib/serrors:go_default_library",
         "//go/lib/snet:go_default_library",
         "//go/lib/sock/reliable:go_default_library",
-        "//go/pkg/sig/sigconfig:go_default_library",
+        "//go/pkg/sig/config:go_default_library",
         "//go/sig/internal/snetmigrate:go_default_library",
     ],
 )
